requirements.txt
Flask
Werkzeug
pycryptodome
Pillow
numpy
matplotlib
python-docx
PyPDF2
openpyxl
xlrd
gunicorn
python-dotenv